backend/testProject/requirements.txt
Django==3.0.2
django-cors-headers==3.2.1
django-faker==0.1
djangorestframework==3.11.0
whitenoise==5.0.1
